divide: chat_onefile.js -> [chat.js + build.js + eventHandlers.js + socketEvents.js]
