+++ views/chat.hbs
@@ -7,13 +7,15 @@
           integrity="sha512-zoJXRvW2gC8Z0Xo3lBbao5+AS3g6YWr5ztKqaicua11xHo+AvE1b0lT9ODgrHTmNUxeCw0Ry4BGRYZfXu70weg=="
           crossorigin="anonymous" referrerpolicy="no-referrer"></script>
   <link href='css/chat.css' rel='stylesheet' />
-  <script src="js/chat.js"></script>
+<!--  <script src="js/chat_onefile.js"></script>-->
+  <script type='module' src="js/chat.js"></script>
   <title>Chat</title>
 </head>
 <body>
+<span id='blockIds' style="display:none;">[{{ user.blockIds }}]</span>
 <div id="main-container">
   <section id="lobby">
-    <h1>[Lobby] (UserId: <span id='userId'>{{ userId }}</span>) 새로고침 금지!!</h1>
+    <h1>[Lobby] (UserId: <span id='userId'>{{ user.id }}</span>) 새로고침 금지!!</h1>
     <ul id="roomList"></ul>
     <div class="input-group">
       <input type="radio" id="public" name="room_type" value="public" checked>
